+++ GonuGame/bin/.gitignore
@@ -1,0 +1,5 @@
+/.DS_Store
+/GameBoard.class
+/PlayerTurn.class
+/PlayingPieces.class
+/application/
