+++ GonuGame/src/application/GameBoard.java
@@ -1,4 +1,4 @@
-package blockadeGames;
+package application;
 
 /**
  * This is Darrah's future graph class
